rename to config.py
